backend/CommentClusteringModelAPI/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
DAWG2-Python==0.9.0
fastapi==0.115.13
filelock==3.18.0
fsspec==2025.5.1
h11==0.16.0
hdbscan==0.8.40
huggingface-hub==0.33.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
llvmlite==0.44.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
nltk==3.9.1
numba==0.61.2
numpy==2.2.6
packaging==25.0
pydantic==2.11.7
pydantic_core==2.33.2
pymorphy3==2.0.4
pymorphy3-dicts-ru==2.4.417150.4580142
pynndescent==0.5.13
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
setuptools==80.9.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.1
tqdm==4.67.1
transformers==4.52.4
typing-inspection==0.4.1
typing_extensions==4.14.0
umap-learn==0.5.7
urllib3==2.5.0
uvicorn==0.34.3
